Working on aggregate data and costs
Close to a usable dataframe for applying cost function and selecting which day is best for what group.  May need to include another column in dfx for n_people so that those values can be evaluated.

Could also create a data set for all possible combinations of choices and costs.